Initial version
Basic functioning mirror video feeder. Works in combination with stf_ios_suport/coordinator

Note that when this is run by the coordinator it will ask for permissions from the firewall to listen on a port despite the port being above the privileged 1024 range. To not get asked one will need to add permissions to mirrorfeed into the firewall.